pyproject.toml
[project]
name = "personal-study-assistent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
]
